pyproject.toml
[tool.poetry]
name = "finsure_api"
version = "0.1.0"
description = ""
authors = ["Nathan Lam <nathan.lam@zoho.com>"]
readme = "README.md"
packages = [{include = "finsure"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
black = "^23.3.0"
djangorestframework = "^3.14.0"
mysqlclient = "^2.2.0"
flake8 = "^6.0.0"
flake8-pyproject = "^1.2.3"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120
